Moved behavior stimulus fixtures
Moved behavior stimulus fixtures to local file and not at top of testing tree.
